+++ main-service/src/main/java/ru/practicum/ewm/mainservice/model/Event.java
@@ -48,7 +48,7 @@ public class Event {
     private Location location;
     private boolean paid;
     @Column(name = "participant_limit")
-    private int participantLimit;
+    private long participantLimit;
     @Column(name = "published_on")
     private LocalDateTime publishedOn;
     @Column(name = "request_moderation")
